pyproject.toml
[tool.poetry]
name = "HM-ScalableRecs"
version = "0.1.0"
description = "Let's build a production-ready recommendation system using the H&M Personalized Fashion dataset."
authors = ["Luong Nguyen Dinh <luongnguyendinhgreen@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.9"
datasets = "^2.21.0"
pandas = "^2.2.2"
loguru = "^0.7.2"
mlflow = "^2.15.1"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
fire = "^0.7.0"
load-dotenv = "^0.1.0"
tqdm = "^4.67.1"
acryl-datahub = "^0.14.1.12"
evidently = "^0.4.40"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.4"
ipywidgets = "^8.1.3"
seaborn = "^0.13.2"
matplotlib = "^3.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
